# apps/backend/crates/orderbook/openapi.yml
openapi: 3.0.3
info:
  version: 0.0.1
  title: Order Book API
servers:
  - description: Mainnet (Prod)
    url: "https://api.cow.fi/mainnet"
  - description: Mainnet (Staging)
    url: "https://barn.api.cow.fi/mainnet"
  - description: Gnosis Chain (Prod)
    url: "https://api.cow.fi/xdai"
  - description: Gnosis Chain (Staging)
    url: "https://barn.api.cow.fi/xdai"
  - description: Arbitrum One (Prod)
    url: "https://api.cow.fi/arbitrum_one"
  - description: Arbitrum One (Staging)
    url: "https://barn.api.cow.fi/arbitrum_one"
  - description: Base (Prod)
    url: "https://api.cow.fi/base"
  - description: Base (Staging)
    url: "https://barn.api.cow.fi/base"
  - description: Avalanche (Prod)
    url: "https://api.cow.fi/avalanche"
  - description: Avalanche (Staging)
    url: "https://barn.api.cow.fi/avalanche"
  - description: Polygon (Prod)
    url: "https://api.cow.fi/polygon"
  - description: Polygon (Staging)
    url: "https://barn.api.cow.fi/polygon"
  - description: Linea (Prod)
    url: "https://api.cow.fi/linea"
  - description: Linea (Staging)
    url: "https://barn.api.cow.fi/linea"
  - description: BNB (Prod)
    url: "https://api.cow.fi/bnb"
  - description: BNB (Staging)
    url: "https://barn.api.cow.fi/bnb"
  - description: Plasma (Prod)
    url: "https://api.cow.fi/plasma"
  - description: Plasma (Staging)
    url: "https://barn.api.cow.fi/plasma"
  - description: Ink (Prod)
    url: "https://api.cow.fi/ink"
  - description: Ink (Staging)
...
paths:
...
  "/api/v1/orders/{UID}/pathviz":
    get:
      operationId: getOrderPathViz
      summary: Route/surplus visualization graph for an order (pathviz).
      description: >
        Returns a `PathVizGraph` describing how the order flowed through the
        solver competition and, once settled, the on-chain venues, plus the
        surplus callout. Quote-stage orders disclose solver names only (no
        venue column). 404 when the pathviz feature is disabled or the order
        is unknown. CoW-hosted chains never serve this endpoint.
      parameters:
        - in: path
          name: UID
          schema:
            $ref: "#/components/schemas/UID"
          required: true
        - in: query
          name: pathVizImage
          required: false
          schema:
            type: boolean
            default: false
          description: Include the rendered base64 SVG as `svgBase64`.
      responses:
        "200":
          description: The visualization graph (and optional rendered SVG).
          content:
            application/json:
              schema:
                type: object
                properties:
                  context:
                    type: string
                    enum:
                      - quotedOnly
                      - executing
                      - traded
                    description: Lifecycle stage the graph describes.
                  settlementTxHash:
                    $ref: "#/components/schemas/TransactionHash"
                  graph:
                    $ref: "#/components/schemas/PathVizGraph"
                  svgBase64:
                    type: string
                    description: Rendered SVG, present only when `pathVizImage=true`.
                  generatedAt:
                    type: string
                    description: ISO-8601 UTC generation timestamp.
                required:
                  - context
                  - graph
                  - generatedAt
        "400":
          description: Invalid order UID.
        "404":
          description: pathviz is disabled or the order was not found.
  "/api/v1/orders/{UID}/pathviz.svg":
    get:
      operationId: getOrderPathVizSvg
      summary: Route/surplus visualization as a raw SVG (pathviz).
      description: >
        Renders the order's `PathVizGraph` to a self-contained SVG served
        under a strict Content-Security-Policy (`default-src 'none'`) with
        `X-Content-Type-Options: nosniff`. Traded orders are immutable and
        served with a long, immutable cache; pre-settlement views use a short
        max-age. Render options are accepted as query parameters mirroring
        `PathVizImageConfig`.
      parameters:
        - in: path
          name: UID
          schema:
            $ref: "#/components/schemas/UID"
          required: true
      responses:
        "200":
          description: The rendered SVG.
          content:
            image/svg+xml:
              schema:
                type: string
        "400":
          description: Invalid order UID or malformed render options.
        "404":
          description: pathviz is disabled or the order was not found.
...
components:
  schemas:
    TransactionHash:
      description: 32 byte digest encoded as a hex with `0x` prefix.
      type: string
      example: "0xd51f28edffcaaa76be4a22f6375ad289272c037f3cc072345676e88d92ced8b5"
...
    UID:
      description: |-
        Unique identifier for the order: 56 bytes encoded as hex with `0x`
        prefix.

        Bytes 0..32 are the order digest, bytes 30..52 the owner address and
        bytes 52..56 the expiry (`validTo`) as a `uint32` unix epoch timestamp.
      type: string
      example: >-
        0xff2e2e54d178997f173266817c1e9ed6fee1a1aae4b43971c53b543cffcc2969845c6f5599fbb25dbdd1b9b013daf85c03f3c63763e4bc4a
...
    PathVizGraph:
      description: >
        Server-rendered route/surplus visualization model. `schemaVersion` is
        1, labeled EXPERIMENTAL for 60 days from first ship, then stable and
        additive-only. All atom amounts are decimal strings. Adds, over the
        Odos surface, the solver-competition column and the surplus callout.
      type: object
      properties:
        schemaVersion:
          type: integer
          example: 1
        nodes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              label:
                type: string
              kind:
                type: string
                enum: [token, solver, venue, overflow]
              column:
                type: integer
                description: Layout column 0..3 (input, solver, venue, output).
              address:
                type: string
            required:
              - id
              - label
              - kind
              - column
        links:
          type: array
          items:
            type: object
            properties:
              from:
                type: string
              to:
                type: string
              kind:
                type: string
                enum: [route, bid, matched]
              amountAtoms:
                type: string
              amountDisplay:
                type: string
              tokenSymbol:
                type: string
            required:
              - from
              - to
              - kind
        solvers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              winner:
                type: boolean
              executedSellAtoms:
                type: string
              executedBuyAtoms:
                type: string
            required:
              - name
              - winner
        surplus:
          type: object
          description: >
            Present once settled. The callout wording is
            "surplus returned vs your signed minimum".
          properties:
            amountAtoms:
              type: string
            tokenSymbol:
              type: string
            amountDisplay:
              type: string
            percent:
              type: number
          required:
            - amountAtoms
            - tokenSymbol
        fee:
          type: object
          properties:
            amountAtoms:
              type: string
            tokenSymbol:
              type: string
            amountDisplay:
              type: string
            bps:
              type: string
          required:
            - amountAtoms
            - tokenSymbol
      required:
        - schemaVersion
        - nodes
        - links
        - solvers
...